+++ tests/test_basic.py
@@ -7,7 +7,7 @@ def test_version():
     """Test that version is defined."""
     from efspurge import __version__
 
-    assert __version__ == "1.7.1"
+    assert __version__ == "1.7.2"
 
 
 def test_imports():